move all the multipatch logfile classes together
